requirements.txt
google-generativeai
fpdf2
beautifulsoup4
requests
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
pydantic
textblob
rake-nltk
nltk
python-multipart
python-dotenv
